+++ CHANGELOG.md
@@ -7,5 +7,5 @@ Changelog
     * Use `ReaderT` stack instead of `State`
     * Add `IndentParserT` in addition to `IndentParser`
     * Remove dependency on `concatenative`
-    * Add `topLevel` function
+    * Add `topLevel`, `notTopLevel` functions
     * Add `runIndentParserT`, `runIndentParser` convenience functions